Actual fix for privacy issue
